+++ tests/example/admin.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+
+from .models import Article, Author
+
+
+class ArticleInline(admin.TabularInline):
+    model = Article
+
+
+@admin.register(Author)
+class AuthorAdmin(admin.ModelAdmin):
+    inlines = [
+        ArticleInline,
+    ]