+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ### Unreleased
 
+### 0.4.0
+
 - Add infra for rest framework serializers checks
-- Add checks drf-model-serializer-extra-kwargs, drf-model-serializer-meta-attribute
+- Add checks
+    - drf-model-serializer-extra-kwargs
+    - drf-model-serializer-meta-attribute
+    - model-admin
 
 ### 0.3.0